requirements.txt
pygame
Pillow
pillow-heif
pytest
streamlit
supabase
